+++ Commons/UICore/src/main/java/com/mfino/uicore/fix/processor/impl/CheckBalanceProcessorImpl.java
@@ -87,9 +87,9 @@ public class CheckBalanceProcessorImpl extends BaseFixProcessor implements Check
         	entry.setBalance(ZERO);
         }
         if (null != pocket.getPocketTemplateByPockettemplateid()) {
-            entry.setCommodity(((Long)pocket.getPocketTemplateByPockettemplateid().getCommodity()).intValue());
+            entry.setCommodity((pocket.getPocketTemplateByPockettemplateid().getCommodity()).intValue());
             String PocketTypeText = enumTextService.getEnumTextValue(CmFinoFIX.TagID_PocketType, null, pocket.getPocketTemplateByPockettemplateid().getType());
-            Integer commodityType = ((Long)pocket.getPocketTemplateByPockettemplateid().getCommodity()).intValue();
+            Integer commodityType = (pocket.getPocketTemplateByPockettemplateid().getCommodity()).intValue();
             String commodityText = enumTextService.getEnumTextValue(CmFinoFIX.TagID_Commodity, null, commodityType);
             entry.setPocketTypeText(String.format("%s  %s", commodityText,PocketTypeText));
         }
